+++ quickbooks/client.py
@@ -47,7 +47,7 @@ class QuickBooks(object):
 
     _BUSINESS_OBJECTS = [
         "Account", "Attachable", "Bill", "BillPayment",
-        "Class", "CreditMemo", "Customer", "CompanyCurrency",
+        "Class", "CreditMemo", "Customer", "CustomerType", "CompanyCurrency",
         "Department", "Deposit", "Employee", "Estimate", "ExchangeRate", "Invoice",
         "Item", "JournalEntry", "Payment", "PaymentMethod", "Preferences",
         "Purchase", "PurchaseOrder", "RefundReceipt",