+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Changelog
     * Added new fields to Invoice object
     * Added new fields to Payment object
     * Fixed issue with Paymentline LinkedTxn
+    * Added new object CustomerType
+
 
 * 0.9.1 (November 30, 2021)
     * Added response status code when raising unauthorized exceptions