+++ CHANGELOG.rst
@@ -6,11 +6,11 @@ Changelog
     * Added new fields to CreditCardPayment object
     * Added new fields to Invoice object
     * Added new fields to Payment object
-    * Fixed issue with Paymentline LinkedTxn
+    * Added to_linked_txn method to Payment object
     * Added new object CustomerType
     * Added MetaData to CompanyInfo
     * Added update support to CompanyInfo
-
+    * Added new fields to Preferences object
 
 * 0.9.1 (November 30, 2021)
     * Added response status code when raising unauthorized exceptions